Added night mode + minor bug fix
Now there is a night mode that is activated at power up if the potentiometer is fully closed. It works like this: between 00:00 and 08:00 the time is displayed as usual. Between 8 AM and midnight the display is normally off (except blinking dots, unless it is totally dark) until there is a change in light intensity hitting the LDR (must be installed, obviously). If the light intensity changes more than the threshold when the timeout is about to expire, then the timeout counter is reset to the top value.