Various improvements, see detail
When setting the time/date, if the pot is not rotated the current value is not changed. This way you can cycle to the - say - day-of-month leaving the time intact. Just remember that the time will be written back to :00 seconds.

When setting the month, now the upper limit is enforced, so you cannot try to set 31st February.